requirements.txt
aiortc>=1.10.1
arxiv
beautifulsoup4
chromadb
faiss-cpu
fastapi
langchain
langchain-chroma
langchain-groq
langchain-openai
langchain_huggingface
langchain_community
langchain_core
langchain_text_splitters
numpy>=1.24.3
openai>=1.3.0
pymupdf
pypdf
PyAudio
sentence_transformers
sounddevice
soundfile
python-docx>=0.8.11
streamlit-webrtc>=0.47.9
supabase
torch
transformers
uvicorn
wikipedia
openpyxl>=3.0.10
python-pptx>=0.6.21
pdfplumber>=0.7.5
streamlit-mic-recorder
google-api-python-client>=2.100.0
wikipedia-api>=0.6.0
beautifulsoup4>=4.12.2
requests>=2.31.0
